Generated Latest Changes for v2021-02-25

+++ subscription_ramp_interval_response.go
@@ -19,7 +19,7 @@ type SubscriptionRampIntervalResponse struct {
 	RemainingBillingCycles int `json:"remaining_billing_cycles,omitempty"`
 
 	// Represents the price for the ramp interval.
-	UnitAmount int `json:"unit_amount,omitempty"`
+	UnitAmount float64 `json:"unit_amount,omitempty"`
 }
 
 // GetResponse returns the ResponseMetadata that generated this resource
